+++ src/components/SubNav/subNavChangelogData.ts
@@ -10,10 +10,17 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: ["Added dark mode color styles."],
+  },
+  {
     date: "2024-12-05",
     version: "3.5.0",
     type: "New Feature",
     affects: ["Documentation", "Functionality"],
-    notes: ["Added the `SubNav` component"],
+    notes: ["Added the `SubNav` component."],
   },
 ];